worked on compiler warnings
git-svn-id: svn://svn.code.sf.net/p/pmii/code/trunk@5028 2dc5e0b7-7910-0410-9abd-fb16afba6634

+++ QedeqKernelSeTest/src/org/qedeq/kernel/se/visitor/QedeqNumbersTest.java
@@ -66,7 +66,7 @@ public class QedeqNumbersTest extends QedeqTestCase {
         assertEquals(0, numbers.getPredicateDefinitionNumber());
         assertEquals(0, numbers.getFunctionDefinitionNumber());
         assertEquals(0, numbers.getRuleNumber());
-        assertEquals((double) 0, numbers.getVisitPercentage(), 1e-10);
+        assertEquals(0, numbers.getVisitPercentage(), 1e-10);
     }
 
     /**
